Handle group placements properly when terminating daycare placements
Re-use clearGroupPlacementsAfter logic in placement termination.
Move placement related queries to PlacementQueries.kt
