+++ service/src/main/kotlin/fi/espoo/evaka/placement/PlacementPlanQueries.kt
@@ -9,12 +9,9 @@ import fi.espoo.evaka.application.DaycarePlacementPlan
 import fi.espoo.evaka.shared.ApplicationId
 import fi.espoo.evaka.shared.ChildId
 import fi.espoo.evaka.shared.DaycareId
-import fi.espoo.evaka.shared.EvakaUserId
 import fi.espoo.evaka.shared.PersonId
-import fi.espoo.evaka.shared.PlacementId
 import fi.espoo.evaka.shared.PlacementPlanId
 import fi.espoo.evaka.shared.db.Database
-import fi.espoo.evaka.shared.db.bindNullable
 import fi.espoo.evaka.shared.domain.FiniteDateRange
 import fi.espoo.evaka.shared.domain.NotFound
 import org.jdbi.v3.core.kotlin.mapTo
@@ -257,61 +254,4 @@ fun Database.Read.getGuardiansRestrictedStatus(guardianId: PersonId): Boolean? {
         .mapTo<Boolean>()
         .toList()
         .singleOrNull()
-}
-
-fun Database.Transaction.deleteServiceNeedsFromPlacementAfter(id: PlacementId, date: LocalDate) {
-    createUpdate(
-        """
-DELETE FROM service_need 
-WHERE placement_id = :placementId
-    AND start_date > :date
-        """.trimIndent()
-    )
-        .bind("placementId", id)
-        .bind("date", date)
-        .execute()
-
-    createUpdate(
-        //language=SQL
-        """
-UPDATE service_need 
-SET end_date = :date 
-WHERE placement_id = :placementId
-    AND daterange(start_date, end_date, '[]') @> :date::date
-        """.trimIndent()
-    )
-        .bind("placementId", id)
-        .bind("date", date)
-        .execute()
-}
-
-fun Database.Transaction.terminatePlacementFrom(terminationRequestedDate: LocalDate, placementId: PlacementId, placementTerminationDate: LocalDate, terminatedBy: EvakaUserId?) {
-    createUpdate(
-        //language=SQL
-        """
-UPDATE daycare_group_placement 
-SET end_date = :placementTerminationDate 
-WHERE daycare_placement_id = :placementId
-        """.trimIndent()
-    )
-        .bind("placementId", placementId)
-        .bind("placementTerminationDate", placementTerminationDate)
-        .execute()
-
-    deleteServiceNeedsFromPlacementAfter(placementId, placementTerminationDate)
-
-    createUpdate(
-        """
-UPDATE placement
-SET termination_requested_date = :terminationRequestedDate,
-    terminated_by = :terminationRequestedBy,
-    end_date = :placementTerminationDate
-WHERE id = :placementId
-        """.trimIndent()
-    )
-        .bindNullable("terminationRequestedDate", if (terminatedBy == null) null else terminationRequestedDate)
-        .bindNullable("terminationRequestedBy", terminatedBy)
-        .bind("placementTerminationDate", placementTerminationDate)
-        .bind("placementId", placementId)
-        .execute()
 }
