+++ service/src/main/kotlin/fi/espoo/evaka/placement/PlacementQueries.kt
@@ -6,6 +6,7 @@ package fi.espoo.evaka.placement
 
 import fi.espoo.evaka.shared.ChildId
 import fi.espoo.evaka.shared.DaycareId
+import fi.espoo.evaka.shared.EvakaUserId
 import fi.espoo.evaka.shared.GroupId
 import fi.espoo.evaka.shared.GroupPlacementId
 import fi.espoo.evaka.shared.PlacementId
@@ -194,19 +195,13 @@ fun Database.Transaction.cancelPlacement(id: PlacementId): Triple<ChildId, Local
 
     getPlacement(id) ?: throw NotFound("Placement $id not found")
 
-    createUpdate(
-        //language=SQL
-        """DELETE FROM daycare_group_placement WHERE daycare_placement_id = :id""".trimIndent()
-    )
+    createUpdate("DELETE FROM daycare_group_placement WHERE daycare_placement_id = :id")
         .bind("id", id)
         .execute()
 
     deleteServiceNeedsFromPlacement(id)
 
-    return createQuery(
-        //language=SQL
-        """DELETE FROM placement WHERE id = :id RETURNING child_id, start_date, end_date""".trimIndent()
-    )
+    return createQuery("DELETE FROM placement WHERE id = :id RETURNING child_id, start_date, end_date")
         .bind("id", id)
         .mapTo<QueryResult>()
         .first()
@@ -215,7 +210,6 @@ fun Database.Transaction.cancelPlacement(id: PlacementId): Triple<ChildId, Local
 
 fun Database.Transaction.clearGroupPlacementsAfter(placementId: PlacementId, date: LocalDate) {
     createUpdate(
-        //language=SQL
         """
             DELETE from daycare_group_placement
             WHERE daycare_placement_id = :id AND start_date > :date
@@ -226,7 +220,6 @@ fun Database.Transaction.clearGroupPlacementsAfter(placementId: PlacementId, dat
         .execute()
 
     createUpdate(
-        //language=SQL
         """
             UPDATE daycare_group_placement SET end_date = :date
             WHERE daycare_placement_id = :id AND start_date <= :date AND end_date > :date
@@ -648,4 +641,49 @@ private val toDaycarePlacement: (ResultSet, StatementContext) -> DaycarePlacemen
         endDate = rs.getDate("placement_end").toLocalDate(),
         type = rs.getEnum("placement_type")
     )
+}
+
+fun Database.Transaction.deleteServiceNeedsFromPlacementAfter(id: PlacementId, date: LocalDate) {
+    createUpdate(
+        """
+DELETE FROM service_need
+WHERE placement_id = :placementId
+    AND start_date > :date
+        """.trimIndent()
+    )
+        .bind("placementId", id)
+        .bind("date", date)
+        .execute()
+
+    createUpdate(
+        """
+UPDATE service_need
+SET end_date = :date
+WHERE placement_id = :placementId
+    AND daterange(start_date, end_date, '[]') @> :date::date
+        """.trimIndent()
+    )
+        .bind("placementId", id)
+        .bind("date", date)
+        .execute()
+}
+
+fun Database.Transaction.terminatePlacementFrom(terminationRequestedDate: LocalDate, placementId: PlacementId, terminationDate: LocalDate, terminatedBy: EvakaUserId?) {
+    clearGroupPlacementsAfter(placementId, terminationDate)
+    deleteServiceNeedsFromPlacementAfter(placementId, terminationDate)
+
+    createUpdate(
+        """
+UPDATE placement
+SET termination_requested_date = :terminationRequestedDate,
+    terminated_by = :terminationRequestedBy,
+    end_date = :placementTerminationDate
+WHERE id = :placementId
+        """.trimIndent()
+    )
+        .bindNullable("terminationRequestedDate", if (terminatedBy == null) null else terminationRequestedDate)
+        .bindNullable("terminationRequestedBy", terminatedBy)
+        .bind("placementTerminationDate", terminationDate)
+        .bind("placementId", placementId)
+        .execute()
 }
